requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
diffusers>=0.27.0
transformers>=4.38.0
accelerate>=0.27.0
safetensors>=0.4.0

# Image handling
Pillow>=10.0.0

# Optimization
xformers>=0.0.23        # Optional: memory-efficient attention (NVIDIA only)
bitsandbytes>=0.42.0    # Optional: 8-bit optimizers for fine-tuning

# Fine-tuning (LoRA)
peft>=0.9.0

# Dataset preparation
datasets>=2.17.0

# Utilities
tqdm>=4.66.0
omegaconf>=2.3.0
